#114: Add pregenerationLastTask to ScriptGenerator
