+++ dsl/src/main/kotlin/edu/wpi/axon/dsl/task/CheckpointCallbackTask.kt
@@ -67,7 +67,7 @@ class CheckpointCallbackTask(name: String) : BaseTask(name) {
     override val outputs: Set<Variable>
         get() = setOf(output)
 
-    override val dependencies: Set<Code<*>> = setOf()
+    override val dependencies: MutableSet<Code<*>> = mutableSetOf()
 
     override fun code() = """
         |${output.name} = tf.keras.callbacks.ModelCheckpoint(
