+++ dsl/src/test/kotlin/edu/wpi/axon/dsl/ScriptGeneratorIntegrationTest.kt
@@ -10,9 +10,9 @@ import edu.wpi.axon.testutil.KoinTestFixture
 import io.kotlintest.assertions.arrow.nel.shouldHaveSize
 import io.kotlintest.assertions.arrow.validation.shouldBeInvalid
 import io.kotlintest.assertions.arrow.validation.shouldBeValid
+import io.kotlintest.matchers.string.shouldContainInOrder
 import io.kotlintest.shouldBe
 import java.util.concurrent.CountDownLatch
-import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import org.koin.core.context.startKoin
 
@@ -193,6 +193,8 @@ internal class ScriptGeneratorIntegrationTest : KoinTestFixture() {
                 code = "task1"
             }
             val task2 by tasks.running(MockTask::class) {
+                // This should be generated because it is the lastTask AND generation will happen
+                // because there is a required variable
                 code = "task2"
             }
             lastTask = task2
@@ -200,8 +202,36 @@ internal class ScriptGeneratorIntegrationTest : KoinTestFixture() {
         }
 
         scriptGenerator.code().shouldBeValid { (code) ->
-            assertTrue(code.contains("task1"))
-            assertTrue(code.contains("task2"))
+            code.shouldContainInOrder("task1", "task2")
+        }
+    }
+
+    @Test
+    fun `a pregeneration task runs before a normal task`() {
+        startKoin {
+            modules(defaultModule())
+        }
+
+        val scriptGenerator = ScriptGenerator(
+            DefaultPolymorphicNamedDomainObjectContainer.of(),
+            DefaultPolymorphicNamedDomainObjectContainer.of()
+        ) {
+            val task1 = tasks.run(MockTask::class) {
+                code = "task1"
+            }
+            pregenerationLastTask = task1
+
+            val var1 by variables.creating(MockVariable::class)
+            val task2 = tasks.run(MockTask::class) {
+                code = "task2"
+                outputs += var1
+            }
+            lastTask = task2
+            requireGeneration(var1)
+        }
+
+        scriptGenerator.code().shouldBeValid { (code) ->
+            code.shouldContainInOrder("task1", "task2")
         }
     }
 }
